+++ lib/qonfig/commands/add_option.rb
@@ -20,12 +20,21 @@ module Qonfig
       # @param key [Symbol, String]
       # @param value [Object]
       #
+      # @raise [Qonfig::ArgumentError]
+      # @raise [Qonfig::CoreMethodIntersectionError]
+      #
       # @api private
       # @since 0.1.0
       def initialize(key, value)
-        unless key.is_a?(Symbol) || key.is_a?(String)
-          raise Qonfig::ArgumentError, 'Setting key should be a symbol or a string!'
-        end
+        raise(
+          Qonfig::ArgumentError,
+          'Setting key should be a symbol or a string!'
+        ) unless key.is_a?(Symbol) || key.is_a?(String)
+
+        raise(
+          Qonfig::CoreMethodIntersectionError,
+          "<#{key}> key can not be used since this is a private core method"
+        ) if Qonfig::Settings.intersects_core_method?(key)
 
         @key   = key
         @value = value