+++ lib/qonfig/commands/load_from_env/value_converter.rb
@@ -14,19 +14,19 @@ module Qonfig
     #
     # @api private
     # @since 0.2.0
-    FLOAT_PATTERN   = /\A\d+\.\d+\z/
+    FLOAT_PATTERN = /\A\d+\.\d+\z/
 
     # @return [Regexp]
     #
     # @api private
     # @since 0.2.0
-    TRUE_PATTERN    = /\A(t|true)\z/i
+    TRUE_PATTERN = /\A(t|true)\z/i
 
     # @return [Regexp]
     #
     # @api private
     # @since 0.2.0
-    FALSE_PATTERN   = /\A(f|false)\z/i
+    FALSE_PATTERN = /\A(f|false)\z/i
 
     class << self
       # @param env_data [Hash]