+++ lib/qonfig/data_set/class_builder.rb
@@ -10,7 +10,6 @@ module Qonfig
         # @return [Class<Qonfig::DataSet>]
         #
         # @see Qonfig::DataSet
-        # @see Qonfig::DSL
         #
         # @api private
         # @since 0.2.0